test(parser): isolate legacy DSML compatibility cases
